+++ dConnectDevicePlugin/dConnectDeviceHost/app/src/main/java/org/deviceconnect/android/deviceplugin/host/recorder/HostMediaRecorder.java
@@ -8,7 +8,6 @@ package org.deviceconnect.android.deviceplugin.host.recorder;
 
 import android.content.Context;
 import android.graphics.Rect;
-import android.media.MediaCodecInfo;
 import android.os.Build;
 import android.util.Range;
 import android.util.Size;
@@ -296,214 +295,6 @@ public interface HostMediaRecorder extends HostDevicePhotoRecorder, HostDeviceSt
         }
     }
 
-    enum H264Profile {
-        AVCProfileBaseline("AVCProfileBaseline", MediaCodecInfo.CodecProfileLevel.AVCProfileBaseline),
-        AVCProfileConstrainedBaseline("AVCProfileConstrainedBaseline", MediaCodecInfo.CodecProfileLevel.AVCProfileConstrainedBaseline),
-        AVCProfileConstrainedHigh("AVCProfileConstrainedHigh", MediaCodecInfo.CodecProfileLevel.AVCProfileConstrainedHigh),
-        AVCProfileExtended("AVCProfileExtended", MediaCodecInfo.CodecProfileLevel.AVCProfileExtended),
-        AVCProfileHigh("AVCProfileHigh", MediaCodecInfo.CodecProfileLevel.AVCProfileHigh),
-        AVCProfileHigh10("AVCProfileHigh10", MediaCodecInfo.CodecProfileLevel.AVCProfileHigh10),
-        AVCProfileHigh422("AVCProfileHigh422", MediaCodecInfo.CodecProfileLevel.AVCProfileHigh422),
-        AVCProfileHigh444("AVCProfileHigh444", MediaCodecInfo.CodecProfileLevel.AVCProfileHigh444),
-        AVCProfileMain("AVCProfileMain", MediaCodecInfo.CodecProfileLevel.AVCProfileMain);
-
-        private final String mName;
-        private final int mValue;
-
-        H264Profile(String name, int value) {
-            mName = name;
-            mValue = value;
-        }
-
-        public String getName() {
-            return mName;
-        }
-
-        public int getValue() {
-            return mValue;
-        }
-
-        public static H264Profile nameOf(String name) {
-            for (H264Profile p : values()) {
-                if (p.mName.equalsIgnoreCase(name)) {
-                    return p;
-                }
-            }
-            return null;
-        }
-
-        public static H264Profile valueOf(int value) {
-            for (H264Profile l : values()) {
-                if (l.mValue == value) {
-                    return l;
-                }
-            }
-            return null;
-        }
-    }
-
-    enum H264Level {
-        AVCLevel1("AVCLevel1", MediaCodecInfo.CodecProfileLevel.AVCLevel1),
-        AVCLevel11("AVCLevel1", MediaCodecInfo.CodecProfileLevel.AVCLevel1),
-        AVCLevel12("AVCLevel12", MediaCodecInfo.CodecProfileLevel.AVCLevel12),
-        AVCLevel13("AVCLevel13", MediaCodecInfo.CodecProfileLevel.AVCLevel13),
-        AVCLevel1b("AVCLevel1b", MediaCodecInfo.CodecProfileLevel.AVCLevel1b),
-        AVCLevel2("AVCLevel2", MediaCodecInfo.CodecProfileLevel.AVCLevel2),
-        AVCLevel21("AVCLevel21", MediaCodecInfo.CodecProfileLevel.AVCLevel21),
-        AVCLevel22("AVCLevel22", MediaCodecInfo.CodecProfileLevel.AVCLevel22),
-        AVCLevel3("AVCLevel3", MediaCodecInfo.CodecProfileLevel.AVCLevel3),
-        AVCLevel31("AVCLevel31", MediaCodecInfo.CodecProfileLevel.AVCLevel31),
-        AVCLevel32("AVCLevel32", MediaCodecInfo.CodecProfileLevel.AVCLevel32),
-        AVCLevel4("AVCLevel4", MediaCodecInfo.CodecProfileLevel.AVCLevel4),
-        AVCLevel41("AVCLevel41", MediaCodecInfo.CodecProfileLevel.AVCLevel41),
-        AVCLevel42("AVCLevel42", MediaCodecInfo.CodecProfileLevel.AVCLevel42),
-        AVCLevel5("AVCLevel5", MediaCodecInfo.CodecProfileLevel.AVCLevel5),
-        AVCLevel51("AVCLevel51", MediaCodecInfo.CodecProfileLevel.AVCLevel51),
-        AVCLevel52("AVCLevel52", MediaCodecInfo.CodecProfileLevel.AVCLevel52),
-        AVCLevel6("AVCLevel6", MediaCodecInfo.CodecProfileLevel.AVCLevel6),
-        AVCLevel61("AVCLevel61", MediaCodecInfo.CodecProfileLevel.AVCLevel61),
-        AVCLevel62("AVCLevel62", MediaCodecInfo.CodecProfileLevel.AVCLevel62);
-
-        private final String mName;
-        private final int mValue;
-
-        H264Level(String name, int value) {
-            mName = name;
-            mValue = value;
-        }
-
-        public String getName() {
-            return mName;
-        }
-
-        public int getValue() {
-            return mValue;
-        }
-
-        public static H264Level nameOf(String name) {
-            for (H264Level l : values()) {
-                if (l.mName.equalsIgnoreCase(name)) {
-                    return l;
-                }
-            }
-            return null;
-        }
-
-        public static H264Level valueOf(int value) {
-            for (H264Level l : values()) {
-                if (l.mValue == value) {
-                    return l;
-                }
-            }
-            return null;
-        }
-    }
-
-    enum H265Profile {
-        HEVCProfileMain("HEVCProfileMain", MediaCodecInfo.CodecProfileLevel.HEVCProfileMain),
-        HEVCProfileMain10("HEVCProfileMain10", MediaCodecInfo.CodecProfileLevel.HEVCProfileMain10),
-        HEVCProfileMain10HDR10("HEVCProfileMain10HDR10", MediaCodecInfo.CodecProfileLevel.HEVCProfileMain10HDR10),
-        HEVCProfileMain10HDR10Plus("HEVCProfileMain10HDR10Plus", MediaCodecInfo.CodecProfileLevel.HEVCProfileMain10HDR10Plus),
-        HEVCProfileMainStill("HEVCProfileMainStill", MediaCodecInfo.CodecProfileLevel.HEVCProfileMainStill);
-
-        private final String mName;
-        private final int mValue;
-
-        H265Profile(String name, int value) {
-            mName = name;
-            mValue = value;
-        }
-
-        public String getName() {
-            return mName;
-        }
-
-        public int getValue() {
-            return mValue;
-        }
-
-        public static H265Profile nameOf(String name) {
-            for (H265Profile p : values()) {
-                if (p.mName.equalsIgnoreCase(name)) {
-                    return p;
-                }
-            }
-            return null;
-        }
-
-        public static H265Profile valueOf(int value) {
-            for (H265Profile l : values()) {
-                if (l.mValue == value) {
-                    return l;
-                }
-            }
-            return null;
-        }
-    }
-    
-    enum H265Level {
-        HEVCHighTierLevel1("HEVCHighTierLevel1", MediaCodecInfo.CodecProfileLevel.HEVCHighTierLevel1),
-        HEVCHighTierLevel2("HEVCHighTierLevel2", MediaCodecInfo.CodecProfileLevel.HEVCHighTierLevel2),
-        HEVCHighTierLevel21("HEVCHighTierLevel21", MediaCodecInfo.CodecProfileLevel.HEVCHighTierLevel21),
-        HEVCHighTierLevel3("HEVCHighTierLevel3", MediaCodecInfo.CodecProfileLevel.HEVCHighTierLevel3),
-        HEVCHighTierLevel31("HEVCHighTierLevel31", MediaCodecInfo.CodecProfileLevel.HEVCHighTierLevel31),
-        HEVCHighTierLevel4("HEVCHighTierLevel4", MediaCodecInfo.CodecProfileLevel.HEVCHighTierLevel4),
-        HEVCHighTierLevel41("HEVCHighTierLevel41", MediaCodecInfo.CodecProfileLevel.HEVCHighTierLevel41),
-        HEVCHighTierLevel5("HEVCHighTierLevel5", MediaCodecInfo.CodecProfileLevel.HEVCHighTierLevel5),
-        HEVCHighTierLevel51("HEVCHighTierLevel51", MediaCodecInfo.CodecProfileLevel.HEVCHighTierLevel51),
-        HEVCHighTierLevel52("HEVCHighTierLevel52", MediaCodecInfo.CodecProfileLevel.HEVCHighTierLevel52),
-        HEVCHighTierLevel6("HEVCHighTierLevel6", MediaCodecInfo.CodecProfileLevel.HEVCHighTierLevel6),
-        HEVCHighTierLevel61("HEVCHighTierLevel61", MediaCodecInfo.CodecProfileLevel.HEVCHighTierLevel61),
-        HEVCHighTierLevel62("HEVCHighTierLevel62", MediaCodecInfo.CodecProfileLevel.HEVCHighTierLevel62),
-        HEVCMainTierLevel1("HEVCMainTierLevel1", MediaCodecInfo.CodecProfileLevel.HEVCMainTierLevel1),
-        HEVCMainTierLevel2("HEVCMainTierLevel2", MediaCodecInfo.CodecProfileLevel.HEVCMainTierLevel2),
-        HEVCMainTierLevel21("HEVCMainTierLevel21", MediaCodecInfo.CodecProfileLevel.HEVCMainTierLevel21),
-        HEVCMainTierLevel3("HEVCMainTierLevel3", MediaCodecInfo.CodecProfileLevel.HEVCMainTierLevel3),
-        HEVCMainTierLevel31("HEVCMainTierLevel31", MediaCodecInfo.CodecProfileLevel.HEVCMainTierLevel31),
-        HEVCMainTierLevel4("HEVCMainTierLevel4", MediaCodecInfo.CodecProfileLevel.HEVCMainTierLevel4),
-        HEVCMainTierLevel41("HEVCMainTierLevel41", MediaCodecInfo.CodecProfileLevel.HEVCMainTierLevel41),
-        HEVCMainTierLevel5("HEVCMainTierLevel5", MediaCodecInfo.CodecProfileLevel.HEVCMainTierLevel5),
-        HEVCMainTierLevel51("HEVCMainTierLevel51", MediaCodecInfo.CodecProfileLevel.HEVCMainTierLevel51),
-        HEVCMainTierLevel52("HEVCMainTierLevel52", MediaCodecInfo.CodecProfileLevel.HEVCMainTierLevel52),
-        HEVCMainTierLevel6("HEVCMainTierLevel6", MediaCodecInfo.CodecProfileLevel.HEVCMainTierLevel6),
-        HEVCMainTierLevel61("HEVCMainTierLevel61", MediaCodecInfo.CodecProfileLevel.HEVCMainTierLevel61),
-        HEVCMainTierLevel62("HEVCMainTierLevel62", MediaCodecInfo.CodecProfileLevel.HEVCMainTierLevel62);
-
-        private final String mName;
-        private final int mValue;
-
-        H265Level(String name, int value) {
-            mName = name;
-            mValue = value;
-        }
-
-        public String getName() {
-            return mName;
-        }
-
-        public int getValue() {
-            return mValue;
-        }
-
-        public static H265Level nameOf(String name) {
-            for (H265Level l : values()) {
-                if (l.mName.equalsIgnoreCase(name)) {
-                    return l;
-                }
-            }
-            return null;
-        }
-
-        public static H265Level valueOf(int value) {
-            for (H265Level l : values()) {
-                if (l.mValue == value) {
-                    return l;
-                }
-            }
-            return null;
-        }
-    }
-
     enum BitRateMode {
         VBR("vbr"),
         CBR("cbr"),
@@ -662,10 +453,26 @@ public interface HostMediaRecorder extends HostDevicePhotoRecorder, HostDeviceSt
             mPref.clear();
         }
 
+        /**
+         * 自動フォーカスモードを取得します.
+         *
+         * 未設定の場合は null を返却します。
+         *
+         * @return 自動フォーカスモード
+         */
         public Integer getPreviewAutoFocusMode() {
             return mPref.getInteger("preview_auto_focus", null);
         }
 
+        /**
+         * 自動フォーカスモードを設定します.
+         *
+         * mode が設定された場合には、未設定にします。
+         *
+         * サポートされていないモードが設定された場合には例外が発生します。
+         *
+         * @param mode 自動フォーカスモード
+         */
         public void setPreviewAutoFocusMode(Integer mode) {
             if (mode == null) {
                 mPref.remove("preview_auto_focus");
@@ -689,6 +496,8 @@ public interface HostMediaRecorder extends HostDevicePhotoRecorder, HostDeviceSt
         /**
          * プレビューの配信エンコードの名前を取得します.
          *
+         * 未設定の場合は h264 を返却します。
+         *
          * @return プレビューの配信エンコードの名前
          */
         public String getPreviewEncoder() {
@@ -711,6 +520,13 @@ public interface HostMediaRecorder extends HostDevicePhotoRecorder, HostDeviceSt
             }
         }
 
+        /**
+         * プロファイルとレベルを取得します.
+         *
+         * 未設定の場合には、null を返却します。
+         *
+         * @return プロファイルとレベル
+         */
         public ProfileLevel getProfileLevel() {
             Integer profile = mPref.getInteger("preview_profile", null);
             Integer level = mPref.getInteger("preview_level", null);
@@ -720,6 +536,15 @@ public interface HostMediaRecorder extends HostDevicePhotoRecorder, HostDeviceSt
             return null;
         }
 
+        /**
+         * プロファイルとレベルを設定します.
+         *
+         * null が設定された場合には、未設定にします。
+         *
+         * サポートされていないプロファイルとレベルが設定された場合には例外を発生します。
+         *
+         * @param pl プロファイルとレベル
+         */
         public void setProfileLevel(ProfileLevel pl) {
             if (pl == null) {
                 mPref.remove("preview_profile");
@@ -733,10 +558,20 @@ public interface HostMediaRecorder extends HostDevicePhotoRecorder, HostDeviceSt
             }
         }
 
+        /**
+         * 設定されているプロファイルを取得します.
+         *
+         * @return プロファイル
+         */
         public Integer getProfile() {
             return mPref.getInteger("preview_profile", 0);
         }
 
+        /**
+         * 設定されているレベルを取得します.
+         *
+         * @return レベル
+         */
         public Integer getLevel() {
             return mPref.getInteger("preview_level", 0);
         }
@@ -1036,7 +871,17 @@ public interface HostMediaRecorder extends HostDevicePhotoRecorder, HostDeviceSt
         }
 
         public Float getFocalLength() {
-            return mPref.getFloat("preview_focal_length", null);
+            Float value = mPref.getFloat("preview_focal_length", null);
+            if (value == null) {
+                return null;
+            }
+            List<Float> focalLengthList = getSupportedFocalLengthList();
+            for (Float focalLength : focalLengthList) {
+                if (Math.abs(focalLength - value) < 0.01f) {
+                    return focalLength;
+                }
+            }
+            return null;
         }
 
         public void setFocalLength(Float focalLength) {
@@ -1597,11 +1442,17 @@ public interface HostMediaRecorder extends HostDevicePhotoRecorder, HostDeviceSt
             return max != null && 1.0f <= zoom && zoom <= max;
         }
 
+        /**
+         * 焦点距離をサポートしているか確認します.
+         *
+         * @param focalLength 焦点距離
+         * @return サポートされている場合はtrue、それ以外はfalse
+         */
         public boolean isSupportedFocalLength(Float focalLength) {
             List<Float> list = getSupportedFocalLengthList();
             if (list != null) {
                 for (Float length : list) {
-                    if (length.equals(focalLength)) {
+                    if (Math.abs(focalLength - length) < 0.01f) {
                         return true;
                     }
                 }
